Rename: page, object, layout で分類
