+++ app/style/page/analytics.module.scss
