+++ app/style/page/ranking.module.scss
