+++ app/style/page/trading.module.scss
